requirements_.txt
attrs==20.2.0
Automat==20.2.0
beautifulsoup4==4.9.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
decorator==4.4.2
google-api-core==1.22.2
google-auth==1.21.3
google-cloud-bigquery==1.27.2
google-cloud-core==1.4.1
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
hola==1.1.54
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.5.2
numpy==1.19.2
pandas==1.1.2
parsel==1.6.0
peppercorn==0.6
Protego==0.1.16
protobuf==3.13.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
requests==2.24.0
rsa==4.6
sampleproject==2.0.0
Scrapy==2.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
Twisted==20.3.0
urllib3==1.25.10
validators==0.17.1
w3lib==1.22.0
zope.interface==5.1.0
